ingest-service/pyproject.toml
[tool.poetry]
name = "ingest-service"
version = "0.3.0" # Incrementamos versión por refactorización MiniLM/PyMuPDF
description = "Ingest service for Atenex B2B SaaS (Postgres/Minio/Milvus/SentenceTransformers - CPU - Prefork)"
authors = ["Atenex Team <dev@atenex.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
gunicorn = "^21.2.0"
pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-settings = "^2.2.1"
celery = {extras = ["redis"], version = "^5.3.6"}
# gevent ya no es necesario con prefork
asyncpg = "^0.29.0" # Keep for API
tenacity = "^8.2.3"
python-multipart = "^0.0.9"
structlog = "^24.1.0"
minio = "^7.1.17"

# --- Core Processing Dependencies (REFACTORIZADO) ---
pymupdf = "^1.25.0"                # PyMuPDF for PDF extraction
sentence-transformers = "^2.7"     # For MiniLM embeddings
pymilvus = ">=2.4.1,<2.5.0"         # Official Milvus client

# --- Converter Dependencies (Standalone - MANTENER) ---
# pypdf = ">=4.0.1,<5.0.0"          # ELIMINADO/COMENTADO - Reemplazado por PyMuPDF
python-docx = ">=1.1.0,<2.0.0"
markdown = ">=3.5.1,<4.0.0"
beautifulsoup4 = ">=4.12.3,<5.0.0"
html2text = ">=2024.1.0,<2025.0.0"  # For Markdown -> HTML -> Text conversion

# --- HTTP Client (API - MANTENER) ---
httpx = {extras = ["http2"], version = "^0.27.0"}
h2 = "^4.1.0"

# --- Synchronous DB Dependencies (Worker - MANTENER) ---
sqlalchemy = "^2.0.28"
psycopg2-binary = "^2.9.9"

# --- ELIMINADO ---
# fastembed = ">=0.2.1,<0.3.0"      # ELIMINADO

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"